requirements.txt
fastapi>=0.88.0
pydantic>=1.10.7
mongoengine>=0.24.2
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.1
pymongo~=4.3.3
streamlit~=1.21.0
requests~=2.28.2
pyjwt~=2.6.0
pyevmasm>=0.2.3
jwt>=1.3.1
python-dotenv>=1.0.0